+++ packages/core/src/providers/LocalMulticallProvider.tsx
@@ -1,9 +1,9 @@
 import { ReactNode, useEffect, useState } from 'react'
+import { utils } from 'ethers'
 import { getChainById } from '../helpers'
-import { useEthers } from '../hooks'
-import { useBlockNumber } from './blockNumber'
-import { useConfig, useUpdateConfig } from './config'
+import { useEthers, useBlockNumber, useConfig, useUpdateConfig, useLocalStorage } from '../hooks'
 import multicallABI from '../constants/abi/MultiCall.json'
+import multicall2ABI from '../constants/abi/MultiCall2.json'
 import { deployContract } from '../helpers/contract'
 
 interface LocalMulticallProps {
@@ -20,8 +20,10 @@ enum LocalMulticallState {
 
 export function LocalMulticallProvider({ children }: LocalMulticallProps) {
   const updateConfig = useUpdateConfig()
-  const { multicallAddresses } = useConfig()
   const { library, chainId } = useEthers()
+  const [, setMulticallAddress, getCurrent] = useLocalStorage('local_multicall_address' + chainId)
+
+  const { multicallAddresses, multicallVersion } = useConfig()
   const [localMulticallState, setLocalMulticallState] = useState(LocalMulticallState.Unknown)
   const [multicallBlockNumber, setMulticallBlockNumber] = useState<number>()
   const blockNumber = useBlockNumber()
@@ -34,25 +36,42 @@ export function LocalMulticallProvider({ children }: LocalMulticallProps) {
     } else if (multicallAddresses && multicallAddresses[chainId]) {
       setLocalMulticallState(LocalMulticallState.Deployed)
     } else if (localMulticallState !== LocalMulticallState.Deploying) {
-      const signer = library.getSigner()
-      if (!signer) {
-        setLocalMulticallState(LocalMulticallState.Error)
-        return
-      }
+      const checkDeployed = async () => {
+        const multicallAddress = getCurrent()
 
-      setLocalMulticallState(LocalMulticallState.Deploying)
-
-      const deployMulticall = async () => {
-        try {
-          const { contractAddress, blockNumber } = await deployContract(multicallABI, signer)
-          updateConfig({ multicallAddresses: { [chainId]: contractAddress } })
-          setMulticallBlockNumber(blockNumber)
-          setLocalMulticallState(LocalMulticallState.Deployed)
-        } catch {
-          setLocalMulticallState(LocalMulticallState.Error)
+        if (typeof multicallAddress === 'string' && utils.isAddress(multicallAddress)) {
+          const multicallCode = await library.getCode(multicallAddress)
+          if (multicallCode !== '0x') {
+            updateConfig({ multicallAddresses: { [chainId]: multicallAddress } })
+            return
+          }
         }
+
+        const signer = library && 'getSigner' in library ? library.getSigner() : undefined
+        if (!signer) {
+          setLocalMulticallState(LocalMulticallState.Error)
+          return
+        }
+
+        setLocalMulticallState(LocalMulticallState.Deploying)
+
+        const deployMulticall = async () => {
+          try {
+            const { contractAddress, blockNumber } = await deployContract(
+              multicallVersion === 1 ? multicallABI : multicall2ABI,
+              signer
+            )
+            updateConfig({ multicallAddresses: { [chainId]: contractAddress } })
+            setMulticallAddress(contractAddress)
+            setMulticallBlockNumber(blockNumber)
+            setLocalMulticallState(LocalMulticallState.Deployed)
+          } catch {
+            setLocalMulticallState(LocalMulticallState.Error)
+          }
+        }
+        void deployMulticall()
       }
-      deployMulticall()
+      void checkDeployed()
     }
   }, [library, chainId])