+++ packages/core/src/providers/blockNumber/index.ts
@@ -1,3 +1,2 @@
-export * from './context'
-export * from './provider'
-export * from './reducer'
+export * from './blockNumbers'
+export * from './common'