+++ packages/core/src/providers/transactions/model.ts
@@ -1,5 +1,4 @@
-import { TransactionReceipt, TransactionResponse } from '@ethersproject/providers'
-import { ChainId } from '../../constants'
+import type { TransactionReceipt, TransactionResponse } from '@ethersproject/abstract-provider'
 
 export interface StoredTransaction {
   transaction: TransactionResponse
@@ -10,6 +9,11 @@ export interface StoredTransaction {
   originalTransaction?: TransactionResponse
 }
 
+export type UpdatedTransaction = Omit<StoredTransaction, 'submittedAt'> & { receipt: TransactionReceipt }
+
+/**
+ * @public
+ */
 export function getStoredTransactionState(transaction: StoredTransaction) {
   if (transaction.receipt) {
     return transaction.receipt.status === 0 ? 'Fail' : 'Success'
@@ -18,7 +22,10 @@ export function getStoredTransactionState(transaction: StoredTransaction) {
 }
 
 export type StoredTransactions = {
-  [T in ChainId]?: StoredTransaction[]
+  [chainId: number]: StoredTransaction[]
 }
 
+/**
+ * @internal Intended for internal use - use it on your own risk
+ */
 export const DEFAULT_STORED_TRANSACTIONS: StoredTransactions = {}