+++ packages/core/src/providers/transactions/reducer.ts
@@ -1,35 +1,37 @@
-import { ChainId } from '../../constants'
+import { StoredTransaction, StoredTransactions, UpdatedTransaction } from './model'
 
-import { StoredTransaction, StoredTransactions } from './model'
-
-type Action = AddTransaction | UpdateTransactions
+type Action = AddTransaction | UpdateTransaction | UpdateTransactions
 
 interface AddTransaction {
   type: 'ADD_TRANSACTION'
   payload: StoredTransaction
 }
+interface UpdateTransaction {
+  type: 'UPDATE_TRANSACTION'
+  payload: UpdatedTransaction
+}
 interface UpdateTransactions {
   type: 'UPDATE_TRANSACTIONS'
-  chainId: ChainId
+  chainId: number
   transactions: StoredTransaction[]
-}
-
-function isChainId(chainId: number): chainId is ChainId {
-  return Object.values(ChainId).includes(chainId)
 }
 
 export function transactionReducer(state: StoredTransactions, action: Action): StoredTransactions {
   switch (action.type) {
     case 'ADD_TRANSACTION': {
       const { chainId } = action.payload.transaction
-
-      if (isChainId(chainId)) {
-        return {
-          ...state,
-          [chainId]: [action.payload, ...(state[chainId] ?? [])],
-        }
-      } else {
-        throw TypeError('Unsupported chain')
+      return {
+        ...state,
+        [chainId]: [action.payload, ...(state[chainId] ?? [])],
+      }
+    }
+    case 'UPDATE_TRANSACTION': {
+      const { chainId, hash } = action.payload.transaction
+      return {
+        ...state,
+        [chainId]: (state[chainId] ?? []).map((tx) =>
+          tx.transaction.hash === hash ? { ...tx, ...action.payload } : tx
+        ),
       }
     }
     case 'UPDATE_TRANSACTIONS':