+++ packages/example/src/index.tsx
@@ -1,20 +1,20 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import { Mainnet, DAppProvider } from '@usedapp/core'
-import { App } from './App'
-
-const config = {
-  readOnlyChainId: Mainnet.chainId,
-  readOnlyUrls: {
-    [Mainnet.chainId]: 'https://mainnet.infura.io/v3/3165a249c65f4198bf57200109b8fadf',
-  },
-}
-
-ReactDOM.render(
-  <React.StrictMode>
-    <DAppProvider config={config}>
-      <App />
-    </DAppProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
-)
+export { App } from './App'
+export * from './pages'
+export {
+  Sidebar,
+  SidebarLinkDescription,
+  SidebarNav,
+  SidebarNavLinks,
+  ToMain,
+  ToMainBottom,
+  Handshaking,
+  Page,
+  SidebarContainer,
+  MenuIconContainer,
+  SidebarMenuIconContainer,
+} from './components/base/base'
+export { MenuIcon } from './components/base/MenuIcon'
+export { ENSExample } from './components/ENS/ENSExample'
+export { GlobalStyle } from './global/GlobalStyle'
+export * from './global/styles'
+export { NotificationsList } from './components/Transactions/History'